requirements.txt
absl-py==0.7.1
alembic==1.0.11
astor==0.8.0
blinker==1.4
Click==7.0
Flask==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Mako==1.0.13
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
Pillow==6.1.0
protobuf==3.8.0
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.1
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.5
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
